requirements.txt
anyascii==0.3.3
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
beautifulsoup4==4.14.3
boolean.py==5.0
CacheControl==0.14.4
certifi==2026.1.4
charset-normalizer==3.4.4
cyclonedx-python-lib==11.6.0
defusedxml==0.7.1
Django==5.2.12
django-cors-headers==4.9.0
django-filter==25.2
django-modelcluster==6.4.1
django-modeltranslation==0.20.2
django-permissionedforms==0.1
django-silk==5.5.0
django-stubs-ext==5.2.9
django-taggit==6.1.0
django-tasks==0.9.0
django-treebeard==4.8.0
djangorestframework==3.16.1
draftjs_exporter==5.2.0
drf-spectacular==0.29.0
et_xmlfile==2.0.0
filelock==3.25.1
filetype==1.2.0
gprof2dot==2025.4.14
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
laces==0.1.2
license-expression==30.4.4
markdown-it-py==4.0.0
mdurl==0.1.2
modelsearch==1.1.1
msgpack==1.1.2
openpyxl==3.1.5
packageurl-python==0.17.6
packaging==26.0
pillow==12.1.1
pillow_heif==1.2.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
platformdirs==4.9.4
polib==1.2.0
py-serializable==2.1.0
Pygments==2.19.2
pyparsing==3.3.2
pypdf==6.8.0
python-decouple==3.8
python-magic==0.4.27
python-telegram-bot==21.0
PyYAML==6.0.3
redis==7.3.0
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
sortedcontainers==2.4.0
soupsieve==2.8.3
sqlparse==0.5.5
telepath==0.3.1
tomli==2.4.0
tomli_w==1.2.0
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
wagtail==7.2.3
wagtail-localize==1.13
Willow==1.12.0
